+++ CHANGELOG.md
@@ -9,6 +9,7 @@ __New Features and Enhancements:__
 
 - Add support for folder lock functionality ([#759](https://github.com/box/box-ios-sdk/pull/759))
 - Add support for copyInstanceOnItemCopy field for metadata templates ([#763](https://github.com/box/box-ios-sdk/pull/763))
+- Add support for stream upload of new file versions ([#766](https://github.com/box/box-ios-sdk/pull/766))
 
 __Bug Fixes:__