+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 Changelog
 =========
 
-## Next Release
+## v4.4.0 [2021-04-20]
 
 __Breaking Changes:__
 
 __New Features and Enhancements:__
 
+- Add support for search param to get shared link items ([#756](https://github.com/box/box-ios-sdk/pull/756))
 - Add support for folder lock functionality ([#759](https://github.com/box/box-ios-sdk/pull/759))
 - Add support for copyInstanceOnItemCopy field for metadata templates ([#763](https://github.com/box/box-ios-sdk/pull/763))
 - Add support for stream upload of new file versions and add support for 'If-Match' header when uploading new file versions ([#766](https://github.com/box/box-ios-sdk/pull/766))
@@ -14,6 +15,7 @@ __New Features and Enhancements:__
 
 __Bug Fixes:__
 
+- Pass only a scheme to iOS Authentication APIs ([#755](https://github.com/box/box-ios-sdk/pull/755))
 - Update `listEnterpriseGroups()` to use documented parameter for filtering by name ([#757](https://github.com/box/box-ios-sdk/pull/757))
 - Fix bug for OAuth where the callback is not called if token has been revoked ([#762](https://github.com/box/box-ios-sdk/pull/762))